Scale the the PNGs to [0, 255], instead of normalizing the JPGs to [0, 1].
WEEKLY CHECK IN

What did you accomplish in the past week? What are your goals for the
coming week? Most of my pipeline is working for the object detection
project. Tried DT and linear and RGB SVCs.

Are you currently stuck on anything? If so, what are you facing?
Distribution of pixels seems very different between my training and test
data. I'll elaborate in chat.
Image
Image
Before doing the HSV conversion, I rescale the training images to 0-255
and multiply cast to uint8, since mpimg loads them as [0, 1].
... I'm think that, now that I'm reading this, I remember that mpimg did
something weird with how it loads jpgs (the test images) vs pngs (the
train images)
I see this is discussed under the heading "Make sure your images are
scaled correctly " in teh "Tips and Tricks" section.
huh, well, that doesn't seem to help any. The only difference seems to
be in the scaling, which I already deal with (I think)
See line 148 here:
https://github.com/tsbertalan/vehicleDetection/blob/master/vehicleDetection/features.py#L148
Should the distributions be this different? I could try to deal with it
with some data augmentation, but I'm suspicious that there's a bug
somewhere I should fix instead.
Ah, I think I have it. I was indeed normalizing all images the same, but
to [0, 1], rather than [0, 255]. I think this gave inconsistent results
when I converted them to HSV.
I now instead normalize the the PNGs to [0, 255], instead of normalizing
the JPGs to [0, 1].
The hue distributions are still different, but, looking at this 2d
histogram on a log scale, it's not that there are 0 test images in the H
region where there are lots of training images, and vice/versa:
Image
Download File
Image
Looking at a H colorbar, I'm guessing there are just a lot more trees in
the training data, which is fine:
Image